// src/app/shared/types.ts
import { Method, Role } from "./enums"

export type REQUEST_PARAMS_TYPE<T = any> = {
    url: string
    method: Method
    body?: T
}

export type RESPONSE_TYPE<T = any> = {
    statusCode: number
    message: string
    message_ar?: string
    data: T
}

export type AUTH_BODY_TYPE = {
    role?: Role | string,
    email: string,
    password: string
}

export type AUTH_LOGIN_RESPONSE_TYPE = {
    id: any
    access_token: string
    refresh_token: string
    role: Role
}
export type USER_TYPE = {
    id?: number | string
    name?: string
    email?: string
    role?: Role
}

export type LOG_TYPE = {
    id: number
    user_id: number
    action: string
    timestamp: string
    user: USER_TYPE
}